Rename Cart.js and Catalog.js to match case with classes
