+++ 3-STUDENTS/AndreevValentin/Project/src/components/index.js
@@ -1,5 +1,5 @@
-import Cart from "./cart.js";
-import Catalog from "./catalog.js";
+import Cart from "./Cart.js";
+import Catalog from "./Catalog.js";
 
 export default () => {
 	new Catalog(new Cart(), document.getElementById("featuredItems"), 4);
